Use unsigned long for the ioctl request
This avoids implementation-defined behavior when assigning
FS_IOC_REMOVE_ENCRYPTION_KEY to an int (as there is signed overflow).

The request parameter for ioctl() is also a unsigned long.

+++ fscryptctl.c
@@ -416,7 +416,7 @@ cleanup:
 }
 
 static int cmd_remove_key(int argc, char *const argv[]) {
-  int ioc = FS_IOC_REMOVE_ENCRYPTION_KEY;
+  unsigned long ioc = FS_IOC_REMOVE_ENCRYPTION_KEY;
 
   static const struct option remove_key_options[] = {
       {"all-users", no_argument, NULL, OPT_ALL_USERS}, {NULL, 0, NULL, 0}};
